+++ config/routes.rb
@@ -19,7 +19,6 @@ Thredded::Engine.routes.draw do # rubocop:disable Metrics/BlockLength
         resource :preview, only: [:update], controller: 'private_post_previews'
         member do
           get 'quote'
-          post 'mark_as_unread'
         end
       end
     end
@@ -92,6 +91,13 @@ Thredded::Engine.routes.draw do # rubocop:disable Metrics/BlockLength
   scope path: 'action' do
     # flat urls under here for anything which is non-visible to users & search engines (typically json actions)
     resources :posts, only: %i[] do
+      member do
+        post 'mark_as_read'
+        post 'mark_as_unread'
+      end
+    end
+
+    resources :private_posts, only: %i[] do
       member do
         post 'mark_as_read'
         post 'mark_as_unread'